(refactor) rename meta -> enums
