+++ bot/src/main/java/ru/tinkoff/edu/java/bot/CommandsHandler.java
@@ -4,7 +4,7 @@ import com.pengrad.telegrambot.model.Update;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.tinkoff.edu.java.bot.command.CommandHandler;
-import ru.tinkoff.edu.java.bot.meta.State;
+import ru.tinkoff.edu.java.bot.enums.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.List;
