+++ bot/src/main/java/ru/tinkoff/edu/java/bot/MessageHandler.java
@@ -4,8 +4,7 @@ import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.command.HelpCommandHandler;
-import ru.tinkoff.edu.java.bot.meta.State;
+import ru.tinkoff.edu.java.bot.enums.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.Optional;
