+++ bot/src/main/java/ru/tinkoff/edu/java/bot/UpdatesHandler.java
@@ -6,7 +6,7 @@ import com.pengrad.telegrambot.model.Update;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.meta.State;
+import ru.tinkoff.edu.java.bot.enums.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.HashMap;
