+++ bot/src/test/java/ru/tinkoff/edu/java/bot/command/UpdatesHandlerTest.java
@@ -18,7 +18,7 @@ import ru.tinkoff.edu.java.bot.CommandsHandler;
 import ru.tinkoff.edu.java.bot.MessageHandler;
 import ru.tinkoff.edu.java.bot.UpdatesHandler;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
-import ru.tinkoff.edu.java.bot.meta.State;
+import ru.tinkoff.edu.java.bot.enums.State;
 
 import java.lang.reflect.Field;
 import java.util.List;
